% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{automine}
\alias{automine}
\title{CARET Autominer}
\usage{
automine(formula, model, tactic, data, seed)
}
\arguments{
\item{formula}{Either a plain text formula or an item of the class formula}

\item{model}{CARET model for use, can accept a list through the lapply or sapply wrapper}

\item{tactic}{A string representing the assessment criteria for model, Regression supports 'RMSE' or 'R2' and Classification features 'Kappa' or 'Accuracy'}

\item{data}{A dataframe featuring all the variables used in the formula}

\item{seed}{A random number seed}
}
\description{
This process aims to streamline the process of using CARET for data mining by providing automatic
training test splits as well implementation of models in parallel for cross method comparisons
}
\examples{
Regression
automine(Sepal.Width~., 'earth', 'RMSE', iris, 11)

Classification
automine(Species~., 'rf', 'Kappa', iris, 11)

Multiple Models
modlist<-list('glm','rf','earth')
reg.grid<-sapply(modlist, function(modlist) automine(Sepal.Width~., modlist, 'RMSE', iris, 11)
}
\keyword{caret,}
\keyword{classification,}
\keyword{data}
\keyword{mining,}
\keyword{regression}

